requirements.txt
asgiref==3.4.1
audioread==2.1.9
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
dask==2021.10.0
decorator==5.1.0
Django==3.2.8
django-cors-headers==3.10.0
django-uwsgi==0.2.2
filelock==3.3.0
fonttools==4.25.0
fsspec @ file:///tmp/build/80754af9/fsspec_1636116461911/work
imagecodecs @ file:///tmp/build/80754af9/imagecodecs_1635529108216/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
joblib==1.0.1
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
librosa==0.8.1
llvmlite==0.37.0
locket==0.2.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1638289681807/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
munkres==1.1.4
mysqlclient==2.0.3
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
numba==0.54.0
numpy==1.20.3
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
partd @ file:///tmp/build/80754af9/partd_1618000087440/work
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.4.0
pooch==1.5.1
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1607645421828/work
PyYAML==6.0
resampy==0.2.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
SoundFile==0.10.3.post1
sqlparse==0.4.2
threadpoolctl==2.2.0
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
uWSGI==2.0.20
virtualenv==20.8.1
virtualenv-clone==0.5.7
